Explicit delete config file

+++ src/cmd/uninstall.rs
@@ -12,12 +12,14 @@ Usage:
     pack uninstall -h | --help
 
 Options:
+    -a, --all               Also remove config files
     -h, --help              Display this message
 ";
 
 #[derive(Debug, RustcDecodable)]
 struct UninstallArgs {
     arg_plugin: Vec<String>,
+    flag_all: bool,
 }
 
 pub fn execute(args: &[String]) {
@@ -27,16 +29,16 @@ pub fn execute(args: &[String]) {
     let args: UninstallArgs =
         Docopt::new(USAGE).and_then(|d| d.argv(argv).decode()).unwrap_or_else(|e| e.exit());
 
-    if let Err(e) = uninstall_plugins(args.arg_plugin) {
+    if let Err(e) = uninstall_plugins(args.arg_plugin, args.flag_all) {
         die!("{}", e);
     }
 }
 
-fn uninstall_plugins(plugins: Vec<String>) -> Result<()> {
+fn uninstall_plugins(plugins: Vec<String>, all: bool) -> Result<()> {
     let mut packs = package::fetch()?;
 
     for pack in packs.iter().filter(|p| plugins.contains(&p.name)) {
-        uninstall_plugin(&pack)?;
+        uninstall_plugin(&pack, all)?;
     }
 
     packs.retain(|x| !plugins.contains(&x.name));
@@ -46,11 +48,11 @@ fn uninstall_plugins(plugins: Vec<String>) -> Result<()> {
     Ok(())
 }
 
-fn uninstall_plugin(plugin: &Package) -> Result<()> {
+fn uninstall_plugin(plugin: &Package, all: bool) -> Result<()> {
     let config_file = plugin.config_path();
     let plugin_path = plugin.path();
 
-    if config_file.is_file() {
+    if config_file.is_file() && all {
         fs::remove_file(&config_file)?;
     }
